+++ src/abi/errors.ts
@@ -11,4 +11,5 @@ export const contractErrors = {
     invalidArgument: { id: 134, message: 'Invalid argument' },
     codeNotFound: { id: 135, message: 'Code of a contract was not found' },
     invalidAddress: { id: 136, message: 'Invalid address' },
+    masterchainNotEnabled: { id: 137, message: 'Masterchain support is not enabled for this contract' },
 }