+++ src/benchmarks/benchmarks.spec.ts
@@ -17,8 +17,8 @@ describe('benchmarks', () => {
 
         // Find gas used
         let gasUsed = tracker.collect().reduce((a, v) => a + v.events.reduce((c, d) => d.$type === 'processed' ? c + d.gasUsed : c, 0n), 0n);
-        expect(gasUsed).toMatchInlineSnapshot(`3041n`);
-        expect(functions.init!.code.toBoc().length).toMatchInlineSnapshot(`287`);
+        expect(gasUsed).toMatchInlineSnapshot(`3493n`);
+        expect(functions.init!.code.toBoc().length).toMatchInlineSnapshot(`321`);
     });
     it('benchmark functions(inline)', async () => {
 
@@ -32,7 +32,7 @@ describe('benchmarks', () => {
 
         // Find gas used
         let gasUsed = tracker.collect().reduce((a, v) => a + v.events.reduce((c, d) => d.$type === 'processed' ? c + d.gasUsed : c, 0n), 0n);
-        expect(gasUsed).toMatchInlineSnapshot(`2910n`);
-        expect(functions.init!.code.toBoc().length).toMatchInlineSnapshot(`280`);
+        expect(gasUsed).toMatchInlineSnapshot(`3362n`);
+        expect(functions.init!.code.toBoc().length).toMatchInlineSnapshot(`314`);
     });
 });